Refactor proposal for "KeyValueCommitting"
This only changes the package uploads as an example.

Embrace the imperative style:
  * we carry a lot of state through the commit steps anyway and immutable maps are actually
    dangerous: we may end up reading from the wrong one etc (has happened).
  * performs much better with transactions that have thousands of outputs (the current monadic commit would
    have quite a lot of overhead for setting all these).

Drop the "Option[DamlStateValue]". This was only for dedup value and the extra complexity for catching the case
where some inputs weren't fetched isn't worth the complexity (tests will catch this anyway).

Return an Iterable[(DamlStateKey, DamlStateValue)]. This way we're clear about the ordering of the outputs.