language: cross sdk dalf/dar imports
This adds the possibility to directly import dalfs in a project.
Moreover the imported dalfs/dar can have different sdk versions.

This is done by generating interface files from dalfs and a 'instances'
package that contains the template instance defintions of template data
types. The instances itself are defined via the `external` keyword,
which is replaced by proper daml-lf instance definitions given in the
respective dalf package in the conversion to daml-lf.

We test that cross sdk imports work by importing the `simple-dalf` in
the daml-assistant integation tests.

We changed the default generated migration code to use simple copying of
record fields via `{..}`, but include a separate command to generate
generic instances for a dalf that can be used to define Convertible
instances for deep copies.

We added a test to the assistant integration tests to check that
migrations via generics work.

I'll push commit to update documentation, while this is in review.